+++ test/unit/elastic_class/ElasticSearchDump_init.py
@@ -178,14 +178,23 @@ class UnitTest(unittest.TestCase):
             {"path": {"data": ["/dir/data1"], "logs": ["/dir/logs1"]}}},
             "serverid2": {"name": "hostname2", "settings":
             {"path": {"data": ["/dir/data2"], "logs": ["/dir/logs2"]}}}}
+        self.info_data = {"name": "localservername"}
+        self.health_data = {"status": "green", "cluster_name": "ClusterName"}
+        self.master_name = "MasterName"
+        self.cluster_data = {"_nodes": {"total": 3}}
 
+    @mock.patch("elastic_class.get_cluster_nodes")
+    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.get_cluster_health")
+    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.datetime.datetime")
     @mock.patch("elastic_class.elastic_libs.get_latest_dump")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
     def test_dupe_name(self, mock_es, mock_list, mock_latest, mock_date,
-                       mock_nodes):
+                       mock_nodes, mock_info, mock_health, mock_master,
+                       mock_cluster):
 
         """Function:  test_dupe_name
 
@@ -200,20 +209,28 @@ class UnitTest(unittest.TestCase):
         mock_latest.return_value = self.last_dump
         mock_date.strftime.side_effect = ["dump2", "dump3"]
         mock_nodes.return_value = self.nodes_data
+        mock_info.return_value = self.info_data
+        mock_health.return_value = self.health_data
+        mock_master.return_value = self.master_name
+        mock_cluster.return_value = self.cluster_data
 
         es = elastic_class.ElasticSearchDump(self.host_list, repo=self.repo)
-
         self.assertEqual((es.hosts, es.dump_list, es.repo_name,
                           es.last_dump_name),
                          (self.host_list, self.dump_list, self.repo,
                           self.last_dump))
 
+    @mock.patch("elastic_class.get_cluster_nodes")
+    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.get_cluster_health")
+    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elastic_libs.get_latest_dump")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
     def test_repo_not_passed2(self, mock_es, mock_list, mock_latest,
-                              mock_nodes):
+                              mock_nodes, mock_info, mock_health, mock_master,
+                              mock_cluster):
 
         """Function:  test_repo_not_passed2
 
@@ -229,18 +246,26 @@ class UnitTest(unittest.TestCase):
         mock_list.return_value = self.dump_list
         mock_latest.return_value = self.last_dump
         mock_nodes.return_value = self.nodes_data
+        mock_info.return_value = self.info_data
+        mock_health.return_value = self.health_data
+        mock_master.return_value = self.master_name
+        mock_cluster.return_value = self.cluster_data
 
         es = elastic_class.ElasticSearchDump(self.host_list)
-
         self.assertEqual((es.hosts, es.dump_list, es.repo_name,
                           es.last_dump_name), (self.host_list, [], None, None))
 
+    @mock.patch("elastic_class.get_cluster_nodes")
+    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.get_cluster_health")
+    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elastic_libs.get_latest_dump")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
     def test_repo_not_passed(self, mock_es, mock_list, mock_latest,
-                             mock_nodes):
+                             mock_nodes, mock_info, mock_health, mock_master,
+                             mock_cluster):
 
         """Function:  test_repo_not_passed
 
@@ -254,20 +279,28 @@ class UnitTest(unittest.TestCase):
         mock_list.return_value = self.dump_list
         mock_latest.return_value = self.last_dump
         mock_nodes.return_value = self.nodes_data
+        mock_info.return_value = self.info_data
+        mock_health.return_value = self.health_data
+        mock_master.return_value = self.master_name
+        mock_cluster.return_value = self.cluster_data
 
         es = elastic_class.ElasticSearchDump(self.host_list)
-
         self.assertEqual((es.hosts, es.dump_list, es.repo_name,
                           es.last_dump_name),
                          (self.host_list, self.dump_list, self.repo,
                           self.last_dump))
 
+    @mock.patch("elastic_class.get_cluster_nodes")
+    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.get_cluster_health")
+    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elastic_libs.get_latest_dump")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
     def test_repo_not_present(self, mock_es, mock_list, mock_latest,
-                              mock_nodes):
+                              mock_nodes, mock_info, mock_health, mock_master,
+                              mock_cluster):
 
         """Function:  test_repo_not_present
 
@@ -281,17 +314,25 @@ class UnitTest(unittest.TestCase):
         mock_list.return_value = self.dump_list
         mock_latest.return_value = self.last_dump
         mock_nodes.return_value = self.nodes_data
+        mock_info.return_value = self.info_data
+        mock_health.return_value = self.health_data
+        mock_master.return_value = self.master_name
+        mock_cluster.return_value = self.cluster_data
 
         es = elastic_class.ElasticSearchDump(self.host_list, repo=self.repo2)
-
         self.assertEqual((es.hosts, es.dump_list, es.repo_name,
                           es.last_dump_name), (self.host_list, [], None, None))
 
+    @mock.patch("elastic_class.get_cluster_nodes")
+    @mock.patch("elastic_class.get_master_name")
+    @mock.patch("elastic_class.get_cluster_health")
+    @mock.patch("elastic_class.get_info")
     @mock.patch("elastic_class.get_nodes")
     @mock.patch("elastic_class.elastic_libs.get_latest_dump")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_default(self, mock_es, mock_list, mock_latest, mock_nodes):
+    def test_default(self, mock_es, mock_list, mock_latest, mock_nodes,
+                     mock_info, mock_health, mock_master, mock_cluster):
 
         """Function:  test_default
 
@@ -305,9 +346,12 @@ class UnitTest(unittest.TestCase):
         mock_list.return_value = self.dump_list
         mock_latest.return_value = self.last_dump
         mock_nodes.return_value = self.nodes_data
+        mock_info.return_value = self.info_data
+        mock_health.return_value = self.health_data
+        mock_master.return_value = self.master_name
+        mock_cluster.return_value = self.cluster_data
 
         es = elastic_class.ElasticSearchDump(self.host_list, repo=self.repo)
-
         self.assertEqual((es.hosts, es.dump_list, es.repo_name,
                           es.last_dump_name),
                          (self.host_list, self.dump_list, self.repo,