+++ CHANGELOG.md
@@ -8,10 +8,12 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 Breaking Change
 
 ### Changed
+- elastic_class.ElasticSearch.\_\_init\_\_:  Integrated ElasticCluster class into ElasticSearch class.
 - elastic_class.ElasticSearch.\_\_init\_\_:  Integrated Elastic class into ElasticSearch class.
 - elastic_class.ElasticSearch.\_\_init\_\_:  Replaced attribute updates with call to update_status.
 
 ### Added
+- elastic_class.ElasticSearchStatus:  Added class and associated methods.
 - elastic_class.ElasticSearch.update_status:  Update class attributes by querying Elasticsearch.
 - elastic_class.create_snapshot:  Runs a dump of a repository.
 - elastic_class.create_snapshot_repo:  Creates a repository in Elasticsearch cluster.